Conflicts:
	runtime/stream.c

+++ runtime/stream.c
@@ -1655,7 +1655,7 @@ finalize_it:
 
 /* property set methods */
 /* simple ones first */
-DEFpropSetMeth(strm, iMaxFileSize, int)
+DEFpropSetMeth(strm, iMaxFileSize, int64)
 DEFpropSetMeth(strm, iFileNumDigits, int)
 DEFpropSetMeth(strm, tOperationsMode, int)
 DEFpropSetMeth(strm, tOpenMode, mode_t)
